refactor: #118 - post 정렬 로직을 usePostSort 분리
